requirements.txt
python-dotenv>=1.0.0
pandas>=1.3
requests>=2.25
matplotlib>=3.3
pyTelegramBotAPI>=4.5